finished list balances handler?

+++ resources/model_token_attributes.go
@@ -5,7 +5,7 @@
 package resources
 
 type TokenAttributes struct {
-	Decimals int32  `json:"decimals"`
+	Decimals int64  `json:"decimals"`
 	LogoUri  string `json:"logo_uri"`
 	Name     string `json:"name"`
 	Symbol   string `json:"symbol"`
